refactor(FuzzyMatchComparatorEdit): move to new edit modal
